fix: move headway mode rendering to backend
