+++ assets/css/dup/free_text.scss
@@ -131,7 +131,7 @@
 
   &--dark {
     .free-text {
-      color: #000000;
+      color: #171f26;
     }
 
     .free-text__route-pill {
@@ -190,7 +190,7 @@
 
   &--dark {
     .free-text {
-      color: #000000;
+      color: #171f26;
     }
 
     .free-text__route-pill {
